+++ nds-service-client-impl/src/main/java/br/com/abril/nds/service/impl/NotaFiscalServiceImpl.java
@@ -1740,7 +1740,7 @@ public class NotaFiscalServiceImpl implements NotaFiscalService {
 	}
 
 	@Override
-	public Integer consultaCotaExemplareSumarizadoQtd(
+	public Long consultaCotaExemplareSumarizadoQtd(
 			FiltroViewNotaFiscalDTO filtro) {
 		return this.notaFiscalRepository.consultaCotaExemplaresSumarizadosQtd(filtro);
 	}
